Fixes v0.0.8.1
* Fixed YouTube
* Fixed Shutdown
* Button In asst Fixed
* Fixed paste & open
* Anime
* Minor bugs
