+++ Dockerfile
@@ -13,6 +13,7 @@ RUN git clone https://github.com/TeamUltroid/Ultroid.git /root/TeamUltroid/
 WORKDIR /root/TeamUltroid/
 
 RUN pip3 install --no-cache-dir -r requirements.txt
+
 RUN npm install -g npm@7.16.0 -g
 RUN npm install
 RUN npm run build
