+++ plugins/__init__.py
@@ -30,7 +30,7 @@ except ModuleNotFoundError:
 
 
 start_time = time.time()
-ultroid_version = "v0.0.8"
+ultroid_version = "v0.0.8.1"
 OWNER_NAME = ultroid_bot.me.first_name
 OWNER_ID = ultroid_bot.me.id
