+++ plugins/anime.py
@@ -39,7 +39,6 @@ async def airing_anime(event):
         await event.delete()
 
 
-"""
 @ultroid_cmd(pattern="anime ?(.*)")
 async def anilist(event):
     name = event.pattern_match.group(1)
@@ -50,7 +49,6 @@ async def anilist(event):
     msg = f"**{title}**\n{year} | {episodes} Episodes.\n\n{info}"
     await event.reply(msg, file=banner, link_preview=True)
     await x.delete()
-"""
 
 
 @ultroid_cmd(pattern="character ?(.*)")
