+++ requirements.txt
@@ -1,3 +1,3 @@
-py-Ultroid==2021.6.17
+py-Ultroid==2021.6.18.1
 Shazamio
 https://github.com/New-dev0/Telethon/archive/BETTER.zip
