+++ strings/strings/ka.yml
@@ -1,7 +1,7 @@
 name: Kannada
 natively: ಕನ್ನಡ
 
-author:
+authors:
   - anishgowda21
 
 #commons
